requirements.txt
facenet-pytorch
opencv-python
numpy
requests
rknnlite.api
torch==1.10.1
torchvision==0.11.2